% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{balanced}
\alias{balanced}
\title{Balanced}
\usage{
balanced(f, ...)
}
\arguments{
\item{f}{A function.}

\item{...}{Additional arguments to \code{f} that should \emph{not} be balanced.}
}
\value{
A function like \code{f}, except that missing values are balanced across
\emph{all} inputs when \code{na.rm = TRUE} (except those specified in \code{...}). This
is like using \code{\link[=complete.cases]{complete.cases()}} on the inputs of \code{f}.
}
\description{
Makes a function balance the removal of \code{NA}s across multiple input vectors.
}
\note{
Additional arguments passed to \code{...} are evaluated, so the return
function doesn't evaluate these arguments lazily. In most cases these
additional arguments are parameters like \code{elasticity} or switches like
\code{na.rm}, and eager evaluation prevents these from being changed in,
e.g., a loop. Lazy evaluation can be had by passing an anonymous function
that partials out these arguments.
}
\examples{
p1 <- price6[[3]]
p0 <- price6[[2]]
q1 <- quantity6[[3]]
q0 <- quantity6[[2]]

# Balance missing values for a Fisher index

fisher <- balanced(fisher_index)
fisher(p1, p0, q1, replace(q0, 3, NA), na.rm = TRUE)
fisher_index(p1[-3], p0[-3], q1[-3], q0[-3])

# Operators can be combined, but some care may be needed

x <- 1:6
w <- c(1:5, NA)

f <- factor(rep(letters[1:2], each = 3))

grouped(balanced(fisher_mean), na.rm = TRUE)(x, w, group = f)
balanced(grouped(fisher_mean))(x, w, group = f, na.rm = TRUE)

}
\seealso{
Other operators: 
\code{\link{grouped}()},
\code{\link{quantity_index}()}
}
\concept{operators}
