% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpindex-package.R
\docType{package}
\name{gpindex-package}
\alias{gpindex-package}
\alias{_PACKAGE}
\alias{gpindex}
\title{gpindex: Generalized Price and Quantity Indexes}
\description{
A small package for calculating lots of different price indexes, and by extension quantity indexes. Provides tools to build and work with any type of bilateral generalized-mean index (of which most price indexes are), along with a few important indexes that don't belong to the generalized-mean family (e.g., superlative quadratic-mean indexes, GEKS). Implements and extends many of the methods in Balk (2008, ISBN:978-1-107-40496-0), von der Lippe (2001, ISBN:3-8246-0638-0), and the CPI manual (2020, ISBN:978-1-51354-298-0) for bilateral price indexes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marberts.github.io/gpindex/}
  \item \url{https://github.com/marberts/gpindex}
  \item Report bugs at \url{https://github.com/marberts/gpindex/issues}
}

}
\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com} (\href{https://orcid.org/0000-0003-2544-9480}{ORCID}) [copyright holder]

}
\keyword{internal}
