\docType{package}
\name{gpindex}
\alias{gpindex}

\title{Generalized plus Fisher price and quantity indices}

\description{
A small package for calculating lots of different price and quantity indices.
}

\seealso{
\url{https://github.com/marberts/gpindex}
}

\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com}
}
