% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_indexes.R
\name{index_weights}
\alias{index_weights}
\title{Index weights}
\usage{
index_weights(
  type = c("Carli", "Jevons", "Coggeshall", "Dutot", "Laspeyres", "HybridLaspeyres",
    "LloydMoulton", "Palgrave", "Paasche", "HybridPaasche", "Drobisch", "Unnamed",
    "Tornqvist", "Walsh1", "Walsh2", "MarshallEdgeworth", "GearyKhamis", "Vartia1",
    "MontgomeryVartia", "Vartia2", "SatoVartia", "Theil", "Rao", "Lowe", "Young")
)
}
\arguments{
\item{type}{The name of the index. See details for the possible types of
indexes.}
}
\description{
Calculate weights for a variety of different price indexes.
}
\details{
The \code{index_weights()} function returns a function to calculate weights
for a variety of price indexes. Weights for the following types of indexes
can be calculated.
\itemize{
\item Carli / Jevons / Coggeshall
\item Dutot
\item Laspeyres / Lloyd-Moulton
\item Hybrid Laspeyres (for use in a harmonic mean)
\item Paasche / Palgrave
\item Hybrid Paasche (for use in an arithmetic mean)
\item Törnqvist / Unnamed
\item Drobisch
\item Walsh-I (for an arithmetic Walsh index)
\item Walsh-II (for a geometric Walsh index)
\item Marshall-Edgeworth
\item Geary-Khamis
\item Montgomery-Vartia / Vartia-I
\item Sato-Vartia / Vartia-II
\item Theil
\item Rao
\item Lowe
\item Young
}

The weights need not sum to 1, as this normalization isn't always
appropriate (i.e., for the Vartia-I weights).
}
\note{
Naming for the indexes and weights generally follows the CPI manual (2020),
Balk (2008), and Selvanathan and Rao (1994). In several cases two or more
names correspond to the same weights (e.g., Paasche and Palgrave, or
Sato-Vartia and Vartia-II). The calculations are given in the examples.
}
\examples{
p0 <- price6[[2]]
p1 <- price6[[3]]
q0 <- quantity6[[2]]
q1 <- quantity6[[3]]
pb <- price6[[1]]
qb <- quantity6[[1]]

}
\seealso{
\code{\link[=update_weights]{update_weights()}} for price-updating weights.

\code{\link[=quantity_index]{quantity_index()}} to remap the arguments in these functions for a
quantity index.

Other price-indexes: 
\code{\link{geks}()}
}
\concept{price-indexes}
