\docType{data}
\name{naics2017}
\alias{naics2017}

\title{North American Industry Classification System (NAICS) 2017}

\description{
North American Industry Classification System (NAICS) Canada 2017 Version 3.0, published by Statistics Canada.
}

\usage{
naics2017
}

\format{
A dataframe with 3 columns and 2,084 rows.
\describe{
\item{level}{Position in the hierarchy (integer).}
\item{code}{Classification code (character).}
\item{title}{Description for classification code (character).}
}
}

\source{
Statistics Canada, North American Industry Classification System (NAICS) Canada 2017 Version 3.0, September 14, 2018. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.
}
