% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data.R
\docType{data}
\name{price_data}
\alias{price_data}
\alias{price6}
\alias{quantity6}
\title{Sample price/quantity data}
\format{
Each data frame has 6 rows and 5 columns, with each row corresponding
to a product and each column corresponding to a time period.
}
\source{
Balk, B. M. (2008). \emph{Price and Quantity Index Numbers}.
Cambridge University Press.

ILO, IMF, OECD, UNECE, and World Bank. (2004).
\emph{Producer Price Index Manual: Theory and Practice}.
International Monetary Fund.
}
\description{
Prices and quantities for six products over five periods.
}
\note{
Adapted from tables 3.1 and 3.2 in Balk (2008), which were adapted
from tables 19.1 and 19.2 in the PPI manual.
}
\examples{
# Recreate tables 3.4, 3.6, and 3.12 from Balk (2008).

index_formulas <- function(p1, p0, q1, q0) {
  c(
    harmonic_laspeyres = harmonic_index("Laspeyres")(p1, p0, q0),
    geometric_laspeyres = geometric_index("Laspeyres")(p1, p0, q0),
    laspeyres = arithmetic_index("Laspeyres")(p1, p0, q0),
    paasche = harmonic_index("Paasche")(p1, p0, q1),
    geometric_paasche = geometric_index("Paasche")(p1, p0, q1),
    palgrave = arithmetic_index("Palgrave")(p1, p0, q1),
    fisher = fisher_index(p1, p0, q1, q0),
    tornqvist = geometric_index("Tornqvist")(p1, p0, q1, q0),
    marshall_edgeworth =
        arithmetic_index("MarshallEdgeworth")(p1, p0, q1, q0),
    walsh1 = arithmetic_index("Walsh1")(p1, p0, q1, q0),
    vartia2 = geometric_index("Vartia2")(p1, p0, q1, q0),
    vartia1 = geometric_index("Vartia1")(p1, p0, q1, q0),
    stuvel = stuvel_index(2, 2)(p1, p0, q1, q0)
  )
}

res <- t(mapply(index_formulas, price6, price6[1], quantity6, quantity6[1]))
round(res, 4)

}
