% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice_index}
\alias{splice_index}
\title{Splice an index series}
\usage{
splice_index(x, periods = NULL, initial = NULL)
}
\arguments{
\item{x}{A list of equal-length numeric vectors giving the period-over-period
indexes for each window.}

\item{periods}{A vector (usually numeric) used to subscript each element of
\code{x} and give the splice points for each window. The default splices on each
point in the window.}

\item{initial}{A numeric vector giving an initial period-over-period index
series onto which the elements of \code{x} are spliced. The default uses the
first element of \code{x}.}
}
\value{
A numeric vector giving the spliced (fixed-base) index series.
}
\description{
Splice a collection of index series computed over a rolling window into one
index series. Splicing on multiple points combines the results with a
geometric mean.
}
\examples{
# Make an index series over a rolling window
x <- list(c(1.1, 0.9, 1.2), c(0.8, 1.3, 1.4), c(1.3, 1.3, 0.8))

# Mean splice

splice_index(x)

# Movement splice

splice_index(x, 3)

# Window splice

splice_index(x, 1)

}
\seealso{
Other price-indexes: 
\code{\link{geks}()},
\code{\link{index_weights}()},
\code{\link{price_indexes}}
}
\concept{price-indexes}
