% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpindex-package.R
\docType{package}
\name{gpindex-package}
\alias{gpindex}
\alias{gpindex-package}
\title{gpindex: Generalized Price and Quantity Indexes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to build and work with bilateral generalized-mean price indexes (and by extension quantity indexes), and indexes composed of generalized-mean indexes (e.g., superlative quadratic-mean indexes, GEKS). Covers the core mathematical machinery for making bilateral price indexes, computing price relatives, detecting outliers, and decomposing indexes, with wrappers for all common (and many uncommon) index-number formulas. Implements and extends many of the methods in Balk (2008, \doi{10.1017/CBO9780511720758}), von der Lippe (2007, \doi{10.3726/978-3-653-01120-3}), and the CPI manual (2020, \doi{10.5089/9781484354841.069}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marberts.github.io/gpindex/}
  \item \url{https://github.com/marberts/gpindex}
  \item Report bugs at \url{https://github.com/marberts/gpindex/issues}
}

}
\author{
\strong{Maintainer}: Steve Martin \email{marberts@protonmail.com} (\href{https://orcid.org/0000-0003-2544-9480}{ORCID}) [copyright holder]

}
\keyword{internal}
