\docType{data}
\name{price/quantity data}
\alias{price6}
\alias{quantity6}

\title{Sample price/quantity data}

\description{
Prices and quantities for six products over five periods.
}

\usage{
price6
quantity6
}

\format{
Each matrix has 5 rows and 6 columns, with each row corresponding to a time period and each column corresponding to a product.
}

\note{
Recreated from tables 3.1 and 3.2 in Balk (2008), which were recreated from tables 19.1 and 19.2 in the PPI manual.
}

\source{
Balk, B. M. (2008). \emph{Price and Quantity Index Numbers}. Cambridge University Press.

ILO, IMF, OECD, Eurostat, UN, and World Bank. (2004). \emph{Producer Price Index Manual: Theory and Practice}. International Monetary Fund.
}

