% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{scale_weights}
\alias{scale_weights}
\title{Scale weights}
\usage{
scale_weights(x)
}
\arguments{
\item{x}{A strictly positive numeric vector.}
}
\value{
A numeric vector that sums to 1. If there are \code{NA}s in \code{x} then the result
sums 1 to if these values are removed.
}
\description{
Scale a vector of weights so that they sum to 1.
}
\examples{
scale_weights(1:5)

scale_weights(c(1:5, NA))

}
\seealso{
\code{\link[=grouped]{grouped()}} to make this function applicable to grouped data.

Other weights functions: 
\code{\link{factor_weights}()},
\code{\link{transmute_weights}()}
}
\concept{weights functions}
